+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [3.4.1] - WIP
+## [3.4.2] - WIP
 
 ### Added
 - More Quick dictation options and keyboard shortcuts